refactor(hook): centralize declaration attr
