+++ pkgs/standards/autoapi/autoapi/v3/__init__.py
@@ -39,7 +39,8 @@ from .opspec import (
 )
 
 # ── Ctx-only decorators (new surface; replaces legacy opspec.decorators) ───────
-from .decorators import alias_ctx, op_ctx, hook_ctx, schema_ctx, alias, op_alias
+from .decorators import alias_ctx, op_ctx, schema_ctx, alias, op_alias
+from .hook.decorators import hook_ctx
 
 # ── Bindings (model + API orchestration) ───────────────────────────────────────
 from .bindings import (
