+++ pkgs/standards/autoapi/autoapi/v3/config/constants.py
@@ -63,11 +63,11 @@ DEFAULT_HTTP_METHODS: Mapping[str, Tuple[str, ...]] = {
 
 
 # ───────────────────────────────────────────────────────────────────────────────
-# ‼ Column.info["autoapi"] is deprecated and will be removed. 
+# ‼ Column.info["autoapi"] is deprecated and will be removed.
 # ‼ Support is not guaranteed.
 # Column-level configuration keys (Column.info["autoapi"])
 #   See: v3 schema builder & v3 schema.info check(meta, attr, model)
-# 
+#
 # ───────────────────────────────────────────────────────────────────────────────
 
 COL_LEVEL_CFGS: frozenset[str] = frozenset(
@@ -154,6 +154,7 @@ MODEL_LEVEL_CFGS: frozenset[str] = frozenset(
 
 # Other internal attribute names
 AUTOAPI_CUSTOM_OP_ATTR = "__autoapi_custom_op__"  # marker for custom opspec
+HOOK_DECLS_ATTR = "__autoapi_hook_decls__"  # per-function hook declarations
 
 # ───────────────────────────────────────────────────────────────────────────────
 # ‼ Everything is natively transactional now
@@ -212,6 +213,7 @@ __all__ = [
     "AUTOAPI_DEFAULTS_INCLUDE_ATTR",
     "AUTOAPI_DEFAULTS_EXCLUDE_ATTR",
     "AUTOAPI_CUSTOM_OP_ATTR",
+    "HOOK_DECLS_ATTR",
     "AUTOAPI_TX_MODELS_ATTR",
     "CTX_REQUEST_KEY",
     "CTX_DB_KEY",
