+++ pkgs/standards/autoapi/autoapi/v3/hook/__init__.py
@@ -1,0 +1,21 @@
+from ..config.constants import HOOK_DECLS_ATTR
+from .decorators import hook_ctx
+from .types import PHASE, HookPhase, PHASES, Ctx, StepFn, HookPredicate
+from .shortcuts import hook, hookS
+from ._hook import Hook
+from ..hook_spec import HookSpec
+
+__all__ = [
+    "hook_ctx",
+    "HOOK_DECLS_ATTR",
+    "Hook",
+    "PHASE",
+    "HookPhase",
+    "PHASES",
+    "Ctx",
+    "StepFn",
+    "HookPredicate",
+    "hook",
+    "hookS",
+    "HookSpec",
+]
