+++ pkgs/standards/autoapi/autoapi/v3/opspec/types.py
@@ -2,18 +2,10 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from enum import Enum
-from typing import (
-    Any,
-    Awaitable,
-    Callable,
-    Literal,
-    Mapping,
-    Optional,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import Any, Callable, Literal, Mapping, Optional, Tuple, Type, Union
+
+from ..hook.types import PHASE, HookPhase, PHASES, Ctx, StepFn, HookPredicate
+from ..hook_spec import HookSpec as OpHook
 
 # ───────────────────────────────────────────────────────────────────────────────
 # Core aliases & enums
@@ -40,67 +32,6 @@ TargetOp = Literal[
 ]
 
 VerbAliasPolicy = Literal["both", "alias_only", "canonical_only"]  # legacy export
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Runtime phases (align with runtime/executor.py)
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-class PHASE(str, Enum):
-    PRE_TX_BEGIN = "PRE_TX_BEGIN"
-    START_TX = "START_TX"
-    PRE_HANDLER = "PRE_HANDLER"
-    HANDLER = "HANDLER"
-    POST_HANDLER = "POST_HANDLER"
-    PRE_COMMIT = "PRE_COMMIT"
-    END_TX = "END_TX"
-    POST_COMMIT = "POST_COMMIT"
-    POST_RESPONSE = "POST_RESPONSE"
-    ON_ERROR = "ON_ERROR"
-    ON_PRE_TX_BEGIN_ERROR = "ON_PRE_TX_BEGIN_ERROR"
-    ON_START_TX_ERROR = "ON_START_TX_ERROR"
-    ON_PRE_HANDLER_ERROR = "ON_PRE_HANDLER_ERROR"
-    ON_HANDLER_ERROR = "ON_HANDLER_ERROR"
-    ON_POST_HANDLER_ERROR = "ON_POST_HANDLER_ERROR"
-    ON_PRE_COMMIT_ERROR = "ON_PRE_COMMIT_ERROR"
-    ON_END_TX_ERROR = "ON_END_TX_ERROR"
-    ON_POST_COMMIT_ERROR = "ON_POST_COMMIT_ERROR"
-    ON_POST_RESPONSE_ERROR = "ON_POST_RESPONSE_ERROR"
-    ON_ROLLBACK = "ON_ROLLBACK"
-
-
-HookPhase = Literal[
-    "PRE_TX_BEGIN",
-    "START_TX",
-    "PRE_HANDLER",
-    "HANDLER",
-    "POST_HANDLER",
-    "PRE_COMMIT",
-    "END_TX",
-    "POST_COMMIT",
-    "POST_RESPONSE",
-    "ON_ERROR",
-    "ON_PRE_TX_BEGIN_ERROR",
-    "ON_START_TX_ERROR",
-    "ON_PRE_HANDLER_ERROR",
-    "ON_HANDLER_ERROR",
-    "ON_POST_HANDLER_ERROR",
-    "ON_PRE_COMMIT_ERROR",
-    "ON_END_TX_ERROR",
-    "ON_POST_COMMIT_ERROR",
-    "ON_POST_RESPONSE_ERROR",
-    "ON_ROLLBACK",
-]
-
-PHASES: Tuple[HookPhase, ...] = tuple(p.value for p in PHASE)
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Hook/handler function types
-# ───────────────────────────────────────────────────────────────────────────────
-
-Ctx = Any
-StepFn = Callable[[Ctx], Union[Any, Awaitable[Any]]]
-HookPredicate = Callable[[Any], bool]
 
 # ───────────────────────────────────────────────────────────────────────────────
 # Lazy-capable schema argument types
@@ -143,21 +74,6 @@ SchemaArg = Union[
 DBSpec = Union[str, Mapping[str, Any]]
 
 
-# ───────────────────────────────────────────────────────────────────────────────
-# Hook & Spec dataclasses
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-@dataclass(frozen=True, slots=True)
-class OpHook:
-    phase: HookPhase
-    fn: StepFn
-    order: int = 0
-    when: Optional[HookPredicate] = None
-    name: Optional[str] = None
-    description: Optional[str] = None
-
-
 @dataclass(frozen=True, slots=True)
 class OpSpec:
     """
@@ -185,7 +101,6 @@ class OpSpec:
 
     # Optional per-op engine binding (DSN string or mapping spec)
     db: Optional[DBSpec] = None
-
 
     # HTTP behavior
     arity: Arity = "collection"
